Refactor: Use u8 for NavTarget
I so far have mostly used usize if the systems using it require usize. As far as I know, there is no benefit of using usize here and u8 is appropriate.

+++ src/characters/nav.rs
@@ -50,9 +50,9 @@ pub(crate) struct NavTargetPosMap(HashMap<Entity, Vec2>);
 
 /// Navigation target
 ///
-/// The [`usize`] field is meant as a priority. The higher it is, the more preferred the target is.
+/// [`NavTarget::0`] is meant as a priority. The higher it is, the more preferred the target is.
 #[derive(Component, Default)]
-pub(crate) struct NavTarget(pub(crate) usize);
+pub(crate) struct NavTarget(pub(crate) u8);
 
 /// Navigator that will pathfind to [`NavTarget`]
 #[derive(Component)]
